examples/rag/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
openai>=1.50.0
langfuse>=2.50.0
sentence-transformers>=3.0.0
pydantic>=2.0.0
